refactoring sort_keys and reusing getCh&title to change value
